Initial validation profile implementation
Created initial outline of a field validator that uses a JSON validation profile to determine how to validate the contents of the field. Proof of concept code for validating basic properties of a string field, like min/max and whether the field is required, are implemented. Additional unit tests are required.